+++ packages/plugin-video/web/src/toolbar/videoModal.js
@@ -2,19 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import VideoSelectionInputModal from './videoSelectionInputModal';
 import MediaURLInputModal from './mediaURLInputModal';
+import NewVideoModal from './NewVideoModal';
 import { mediaTypes } from '../types';
-import { createMediaUploadWrapper } from 'wix-rich-content-plugin-commons';
 
 const VideoModal = props => {
   const {
     componentData: { type },
+    experiments,
   } = props;
-  const Comp = mediaTypes.includes(type) ? MediaURLInputModal : VideoSelectionInputModal;
-  return <Comp {...props} />;
+
+  const useNewModal = experiments?.newVideoModal?.enabled;
+  const oldVideoModal = mediaTypes.includes(type) ? MediaURLInputModal : VideoSelectionInputModal;
+  const Component = useNewModal ? NewVideoModal : oldVideoModal;
+
+  return <Component {...props} />;
 };
 
 export default VideoModal;
 
 VideoModal.propTypes = {
   componentData: PropTypes.object,
+  experiments: PropTypes.object,
 };