+++ packages/ricos-content/web/src/normalization/Fixtures/textWrapContentState.ts
@@ -269,7 +269,6 @@ export default {
           alignment: 'center',
           size: 'large',
           width: 'full-width',
-          enableVoteRole: false,
         },
       },
     },