+++ packages/ricos-content/web/src/velo-adapter/adapter.ts
@@ -1,0 +1,2 @@
+export { createBuilder } from './builder';
+export { modify } from './modify';