+++ packages/ricos-content/web/src/velo-adapter/builder.ts
@@ -2,9 +2,9 @@ import { setupContentBuilder } from '../RicosContentAPI/RicosContentBuilder';
 import { generateId } from '../converters/generateRandomId';
 import { flow } from 'lodash';
 import { pipe } from 'fp-ts/function';
-import { applyContent, emptyContent } from './consts';
+import { applyContent } from './consts';
 import { RichContent } from 'ricos-schema';
-import { ContentBuilderAdapter } from './types';
+import { ContentBuilderAdapter, CreateBuilder } from './types';
 import { ContentBuilder } from '../types';
 import { addImage } from './addImage';
 
@@ -12,23 +12,31 @@ const extractBuilderMethods: (
   builder: ReturnType<typeof setupContentBuilder>
 ) => ContentBuilder = ({ RicosContentBuilder: _, ...rest }) => rest;
 
+const notify: CreateBuilder = (content, callback) => {
+  callback?.(content);
+  return createBuilder(content, callback);
+};
+
 const applyBuilderPattern: (
-  content: RichContent
-) => (methods: ContentBuilder) => ContentBuilderAdapter = content => methods =>
+  content: RichContent,
+  callback?: (content: RichContent) => void
+) => (methods: ContentBuilder) => ContentBuilderAdapter = (content, callback) => methods =>
   Object.entries(methods).reduce(
     (prev, [key, func]: [string, (arg) => RichContent]) => ({
       ...prev,
-      [key]: arg => createBuilder(func({ ...arg, content })),
+      [key]: flow(applyContent(content), func, newContent => notify(newContent, callback)),
     }),
     {} as ContentBuilderAdapter
   );
 
-export function createBuilder(content: RichContent = emptyContent): ContentBuilderAdapter {
+export const createBuilder: CreateBuilder = (content, callback?) => {
   const originalMethods = pipe(generateId, setupContentBuilder, extractBuilderMethods);
-  const adapterMethods = pipe(originalMethods, applyBuilderPattern(content));
+  const adapterMethods = pipe(originalMethods, applyBuilderPattern(content, callback));
   return {
     ...adapterMethods,
-    addImage: flow(addImage, applyContent(content), originalMethods.addImage, createBuilder),
+    addImage: flow(addImage, applyContent(content), originalMethods.addImage, newContent =>
+      notify(newContent, callback)
+    ),
     get: () => content,
   };
-}
+};