+++ packages/ricos-editor/web/src/toolbars/TextFormattingToolbar.tsx
@@ -80,22 +80,23 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
 
     const filteredFormattingToolbarButtons = filterButtons(formattingToolbarButtons, activeEditor);
 
+    const getPluginConfig = pluginType =>
+      this.props.plugins?.find(plugin => plugin.type === pluginType)?.config;
+
     const colorPickerData = {
-      TEXT_COLOR: this.props.plugins?.find(plugin => plugin.type === 'wix-rich-content-text-color')
-        ?.config,
-      TEXT_HIGHLIGHT: this.props.plugins?.find(
-        plugin => plugin.type === 'wix-rich-content-text-highlight'
-      )?.config,
+      TEXT_COLOR: getPluginConfig('wix-rich-content-text-color'),
+      TEXT_HIGHLIGHT: getPluginConfig('wix-rich-content-text-highlight'),
     };
     const linkPanelData = {
-      linkTypes: this.props.plugins?.find(plugin => plugin.type === 'LINK')?.config.linkTypes,
+      linkTypes: getPluginConfig('LINK')?.linkTypes,
       uiSettings: { linkPanel: this.props.linkPanelSettings },
       linkSettings: this.props.linkSettings,
       isMobile,
     };
+    const defaultLineSpacing = getPluginConfig('line-spacing')?.defaultSpacing;
+
     const headingsData = {
-      ...this.props.plugins?.find(plugin => plugin.type === 'wix-rich-content-plugin-headings')
-        ?.config,
+      ...getPluginConfig('wix-rich-content-plugin-headings'),
     };
     const onInlineToolbarOpen = () => this.props.onInlineToolbarOpen?.(ToolbarType.FORMATTING);
     const onToolbarButtonClick = (name, value = undefined, pluginId = undefined) => {
@@ -114,6 +115,7 @@ class TextFormattingToolbar extends Component<TextFormattingToolbarProps> {
         headingsData={headingsData}
         onToolbarButtonClick={onToolbarButtonClick}
         experiments={experiments}
+        defaultLineSpacing={defaultLineSpacing}
       />
     );
     const ToolbarContainer =